modified store size pretty print

+++ store.go
@@ -71,9 +71,9 @@ func (st *Store) Size() (int64, string) {
 	}
 	switch {
 	case size > int64(1<<20):
-		return size, fmt.Sprintf("%d MB", size/int64(1<<20))
+		return size, fmt.Sprintf("%.2f MB", float64(size)/float64(1<<20))
 	case size > int64(1<<10):
-		return size, fmt.Sprintf("%d KB", size/int64(1<<10))
+		return size, fmt.Sprintf("%.2f KB", float64(size)/float64(1<<10))
 	}
 	return size, fmt.Sprintf("%d B", size)
 }
